delete 작업중2

+++ index.py
@@ -164,13 +164,25 @@ def post_get():
     except (jwt.ExpiredSignatureError, jwt.exceptions.DecodeError):
         return redirect(url_for("home"))
 
-@app.route("/delete", methods=["POST"])
+
+@app.route("/post/delete", methods=["POST"])
 def post_delete():
+    token_receive = request.cookies.get('mytoken')
+    try:
+        payload = jwt.decode(token_receive, SECRET_KEY, algorithms=['HS256'])
 
-    num_receive = request.form['num_give']
-    db.post.delete_one({'num': int(num_receive)})
-    return jsonify({'msg': '삭제 완료!'})
+        user_info = db.user.find_one({"username": payload["id"]})
+        num_receive = request.form['num_give']
+        poster_info = db.post.find_one({"username": payload["id"]})
 
+        if user_info == poster_info:
+            db.post.delete_one({'num': int(num_receive)})
+            return jsonify({'result': 'success', 'msg': '성공'})
+        else:
+            return jsonify({'result': 'false', 'msg': '실패'})
+
+    except (jwt.ExpiredSignatureError, jwt.exceptions.DecodeError):
+        return redirect(url_for("home"))
 
 
 if __name__ == '__main__':
